fix bug of script. fix bug of config. add simple clean up script.

+++ syncer-core/src/test/java/com/github/zzt93/syncer/config/code/Base.java
@@ -26,7 +26,7 @@ public class Base implements MethodFilter {
         sync.addExtra("suffix", "-" + ((int) sync.getField("tinyint"))/128);
         break;
       case "correctness":
-        sync.updateField("type", Byte.toUnsignedInt((Byte) sync.getField("type")));
+        sync.updateField("type", Byte.toUnsignedInt((byte)(int) sync.getField("type")));
         break;
     }
   }
